+++ src/esm/index.js
@@ -187,11 +187,7 @@ export { StaticTooltip } from './components/tooltips/static-tooltip';
 export { TooltipNew } from './components/tooltips/tooltip-new';
 export { Tooltip } from './components/tooltips/tooltip'; // Tours
 
-export { HighlightHalo } from './components/tours/highlight-halo';
-export { HeaderTour } from './components/tours/header-tour';
-export { TourPopover } from './components/tours/tour-popover';
-export { TourStep } from './components/tours/tour-step';
-export { Tour } from './components/tours/tour'; // Typography
+export { HighlightHalo } from './components/tours/highlight-halo'; // Typography
 
 export { Paragraph } from './components/typography/paragraph';
 export { SupTitle } from './components/typography/sup-title';